Rename Quotes directory to MarketData because I think it's a more accurate name
