+++ tests/test_AlpacaWrapper.py
@@ -1,13 +1,13 @@
 import unittest
 from unittest.mock import patch, MagicMock
-from Quotes.AlpacaWrapper import AlpacaWrapper
+from MarketData.AlpacaWrapper import AlpacaWrapper
 from models.quote import Quote
 
 
 class TestAlpacaWrapper(unittest.TestCase):
 
-    @patch("Quotes.AlpacaWrapper.EnvironmentVariable")
-    @patch("Quotes.AlpacaWrapper.requests.get")
+    @patch("MarketData.AlpacaWrapper.EnvironmentVariable")
+    @patch("MarketData.AlpacaWrapper.requests.get")
     def test_fetch_quote_successful(self, mock_get, mock_env):
         # Arrange
         mock_env_instance = MagicMock()
@@ -39,8 +39,8 @@ class TestAlpacaWrapper(unittest.TestCase):
         self.assertEqual(result.bid_price, 180.0)
         self.assertEqual(result.timestamp, "2025-04-04T19:59:59.874606332Z")
 
-    @patch("Quotes.AlpacaWrapper.EnvironmentVariable")
-    @patch("Quotes.AlpacaWrapper.requests.get")
+    @patch("MarketData.AlpacaWrapper.EnvironmentVariable")
+    @patch("MarketData.AlpacaWrapper.requests.get")
     def test_fetch_quote_raises_for_missing_quotes_key(self, mock_get, mock_env):
         # Arrange
         mock_env_instance = MagicMock()
@@ -60,8 +60,8 @@ class TestAlpacaWrapper(unittest.TestCase):
 
         self.assertIn("No latest quote available for symbol", str(context.exception))
 
-    @patch("Quotes.AlpacaWrapper.EnvironmentVariable")
-    @patch("Quotes.AlpacaWrapper.requests.get")
+    @patch("MarketData.AlpacaWrapper.EnvironmentVariable")
+    @patch("MarketData.AlpacaWrapper.requests.get")
     def test_fetch_quote_raises_on_http_error(self, mock_get, mock_env):
         # Arrange
         mock_env_instance = MagicMock()
